added keep, refactored mity call, normalise, util to use classes

+++ mitylib/normalise.py
@@ -2,14 +2,13 @@
 
 import subprocess
 import logging
-from .util import MityUtil
-from scipy.stats import binom
+import os.path
 from math import isinf
+from scipy.stats import binom
 import numpy as np
-import configparser
 import pysam
 import pysam.bcftools
-import os.path
+from mitylib.util import MityUtil
 
 
 # LOGGING
@@ -37,6 +36,7 @@ class Normalise:
         genome,
         output_file=None,
         allsamples=False,
+        keep=False,
         p=P_VAL,
     ):
         self.debug = debug
@@ -45,6 +45,7 @@ class Normalise:
         self.genome = genome
         self.output_file = output_file
         self.allsamples = allsamples
+        self.keep = keep
         self.p = p
 
         self.normalised_vcf_obj = None
@@ -88,9 +89,10 @@ class Normalise:
         self.do_gsort()
         MityUtil.tabix(output_file)
 
-        # remove temporary files
-        os.remove(self.filtered_vcf_name)
-        os.remove(self.normalised_vcf_name)
+        if not self.keep:
+            # remove temporary files
+            os.remove(self.filtered_vcf_name)
+            os.remove(self.normalised_vcf_name)
 
     def run_bcftools_norm(self):
         """
@@ -210,25 +212,14 @@ class Normalise:
 
     def mity_qual(self, AO, DP, p):
         """
-        Compute variant quality
-        :param AO: (int) number of alternative reads
-        :param DP: (int) total read depth
-        :return: (float) phred-scaled quality score
+        Calculate the Phred-scaled quality score for a genetic variant.
 
-        >>> [mity_qual(x,10) for x in [1,2,3,4,5,6,7,8,9,10]]
-        '[37.49, 60.22, 84.78, 110.97, 138.75, 168.16, 199.4, 232.92, 269.9, 296.89]'
-        >>> [mity_qual(x,100) for x in [5,10,15,20,25,30,35,40,45,50]]
-        '[71.87, 156.08, 251.23, 354.34, 463.9, 579.05, 699.21, 824.04, 953.32, 1086.94]'
-        >>> [mity_qual(x,1000) for x in [5,10,15,20,25,30,35,40,45,50]]
-        '[17.84, 50.97, 93.59, 142.89, 197.36, 256.02, 318.25, 383.57, 451.61, 522.1]'
-        >>> [mity_qual(x,10000) for x in [5,10,15,20,25,30,35,40,45,50]]
-        '[0.0, 0.05, 0.74, 3.56, 9.51, 18.73, 31.0, 46.04, 63.57, 83.37]'
-        >>> mity_qual(118,118)
-        '3211.76'
-        >>> mity_qual(119,119)
-        '3220'
-        >>> mity_qual(10000,10000)
-        '3220'
+        Parameters:
+            - AO (int): Number of alternative reads.
+            - DP (int): Total read depth.
+
+        Returns:
+            - float: Phred-scaled quality score.
         """
         q = 0.0
         AO = int(AO)
@@ -248,6 +239,15 @@ class Normalise:
         return q
 
     def add_filter(self, variant):
+        """
+        Adds filter to variant, and filters for samples.
+
+        Parameters:
+            - variant (VariantRecord): Variant to add filters to
+
+        Returns:
+            - VariantRecord: Variant with filters added
+        """
         # filter dictionary
         # each value represents the number of samples that pass each field
         # starts with all samples passing
@@ -391,5 +391,5 @@ class Normalise:
         """
         Run gsort.
         """
-        gsort_cmd = f"gsort {self.filtered_vcf_name} {self.genome} | bgzip -cf > {self.gsorted_name}"
+        gsort_cmd = f"gsort {self.filtered_vcf_name} {self.genome} | bgzip -cf > {self.output_file}"
         subprocess.run(gsort_cmd, shell=True, check=False)
